backend/cv_engine/faiss_service.py
# FAISS Synchronization Stub

try:
    import faiss
except ImportError:
    faiss = None

try:
    import numpy as np
except ImportError:
    np = None

class FAISSIndexManager:
    def __init__(self, dimension=512):
        self.dimension = dimension
        if faiss:
            self.index = faiss.IndexFlatL2(self.dimension)
        else:
            self.index = None
        self.student_ids = []

    def rebuild_index(self, embeddings_dict):
        """
        Rebuild the FAISS index from a dictionary of {student_id: embedding_vector}.
        """
        self.index.reset()
        self.student_ids = []
        if not embeddings_dict:
            return

        for student_id, vector in embeddings_dict.items():
            self.add_embedding(student_id, vector)

    def add_embedding(self, student_id, vector):
        if len(vector) != self.dimension:
            raise ValueError(f"Vector must be {self.dimension} dimensions")
        
        if not np or not self.index:
            return

        vec_np = np.array([vector], dtype=np.float32)
        self.index.add(vec_np)
        self.student_ids.append(student_id)

    def search(self, vector, k=1, distance_threshold=1.0):
        if not self.index or self.index.ntotal == 0 or not np:
            return None
            
        vec_np = np.array([vector], dtype=np.float32)
        distances, indices = self.index.search(vec_np, k)
        
        if distances[0][0] < distance_threshold and indices[0][0] != -1:
            matched_idx = indices[0][0]
            return self.student_ids[matched_idx]
        return None
