backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
redis>=5.0.1
kafka-python>=2.0.2
faiss-cpu>=1.7.4
opencv-python>=4.8.0
numpy>=1.26.0
deep-sort-realtime>=1.3.2
fer>=22.5.1
tensorflow-cpu>=2.14.0
django-cors-headers>=4.3.0
insightface>=0.7.3
onnxruntime>=1.15.1
celery>=5.3.4
